creation composant scene
