+++ dhareshsawLescargot/src/app/scene/scene-routing.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { CombatsPage } from './combats.page';
+import { ScenePage } from './scene.page';
 
 const routes: Routes = [
   {
     path: '',
-    component: CombatsPage
+    component: ScenePage
   }
 ];
 
@@ -14,4 +14,4 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule],
 })
-export class CombatsPageRoutingModule {}
+export class ScenePageRoutingModule {}
