+++ dhareshsawLescargot/src/app/scene/scene.page.html
@@ -1,6 +1,6 @@
 <ion-header>
   <ion-toolbar>
-    <ion-title>combats</ion-title>
+    <ion-title>scene</ion-title>
   </ion-toolbar>
 </ion-header>
